+++ api/src/main/java/jakarta/security/auth/message/MessagePolicy.java
@@ -15,13 +15,12 @@
  * SPDX-License-Identifier: EPL-2.0 OR GPL-2.0 WITH Classpath-exception-2.0
  */
 
-package javax.security.auth.message;
+package jakarta.security.auth.message;
 
-// for @see
-import javax.security.auth.message.config.ClientAuthContext;
-import javax.security.auth.message.config.ServerAuthContext;
-import javax.security.auth.message.module.ClientAuthModule;
-import javax.security.auth.message.module.ServerAuthModule;
+import jakarta.security.auth.message.config.ClientAuthContext;
+import jakarta.security.auth.message.config.ServerAuthContext;
+import jakarta.security.auth.message.module.ClientAuthModule;
+import jakarta.security.auth.message.module.ServerAuthModule;
 
 /**
  * This class defines a message authentication policy.
