+++ api/src/main/java/jakarta/security/auth/message/ServerAuth.java
@@ -15,7 +15,7 @@
  * SPDX-License-Identifier: EPL-2.0 OR GPL-2.0 WITH Classpath-exception-2.0
  */
 
-package javax.security.auth.message;
+package jakarta.security.auth.message;
 
 import javax.security.auth.Subject;
