+++ api/src/main/java/jakarta/security/auth/message/callback/PrivateKeyCallback.java
@@ -15,7 +15,7 @@
  * SPDX-License-Identifier: EPL-2.0 OR GPL-2.0 WITH Classpath-exception-2.0
  */
 
-package javax.security.auth.message.callback;
+package jakarta.security.auth.message.callback;
 
 import java.math.BigInteger;
 import java.security.PrivateKey;
