+++ api/src/main/java/jakarta/security/auth/message/callback/TrustStoreCallback.java
@@ -15,39 +15,37 @@
  * SPDX-License-Identifier: EPL-2.0 OR GPL-2.0 WITH Classpath-exception-2.0
  */
 
-package javax.security.auth.message.callback;
+package jakarta.security.auth.message.callback;
 
-import java.security.cert.CertStore;
+import java.security.KeyStore;
 import javax.security.auth.callback.Callback;
 
 /**
- * Callback for CertStore.
+ * Callback for trusted certificate KeyStore.
  *
  * <p>
- * A CertStore is a generic repository for certificates. CertStores may be searched to locate public key certificates,
- * as well as to put together certificate chains. Such a search may be necessary when the caller needs to verify a
- * signature.
+ * A trusted certificate KeyStore may be used to determine whether a given certificate chain can be trusted.
  *
  */
-public class CertStoreCallback implements Callback {
+public class TrustStoreCallback implements Callback {
 
-	private CertStore certStore;
+	private KeyStore trustStore;
 
 	/**
-	 * Used by the CallbackHandler to set the CertStore within the Callback.
+	 * Used by the CallbackHandler to set the trusted certificate keystore within the Callback.
 	 *
-	 * @param certStore The certificate store, which may be null
+	 * @param trustStore The trusted certificate KeyStore, which must already be loaded.
 	 */
-	public void setCertStore(CertStore certStore) {
-		this.certStore = certStore;
+	public void setTrustStore(KeyStore trustStore) {
+		this.trustStore = trustStore;
 	}
 
 	/**
-	 * Used by the CertStore user to obtain the CertStore set within the Callback.
+	 * Used by the TrustStore user to obtain the TrustStore set within the Callback.
 	 *
-	 * @return The CertStore, or null.
+	 * @return The trusted certificate KeyStore. The KeyStore is guaranteed to already be loaded.
 	 */
-	public CertStore getCertStore() {
-		return certStore;
+	public KeyStore getTrustStore() {
+		return trustStore;
 	}
 }
