+++ api/src/main/java/jakarta/security/auth/message/config/AuthConfigProvider.java
@@ -15,10 +15,11 @@
  * SPDX-License-Identifier: EPL-2.0 OR GPL-2.0 WITH Classpath-exception-2.0
  */
 
-package javax.security.auth.message.config;
+package jakarta.security.auth.message.config;
 
 import javax.security.auth.callback.CallbackHandler;
-import javax.security.auth.message.AuthException;
+
+import jakarta.security.auth.message.AuthException;
 
 /**
  * This interface is implemented by objects that can be used to obtain authentication context configuration objects,
