+++ api/src/main/java/jakarta/security/auth/message/config/RegistrationListener.java
@@ -15,7 +15,7 @@
  * SPDX-License-Identifier: EPL-2.0 OR GPL-2.0 WITH Classpath-exception-2.0
  */
 
-package javax.security.auth.message.config;
+package jakarta.security.auth.message.config;
 
 /**
  * An implementation of this interface may be associated with an AuthConfigProvider registration at an AuthConfigFactory
