+++ api/src/main/java/jakarta/security/auth/message/config/ServerAuthConfig.java
@@ -15,11 +15,12 @@
  * SPDX-License-Identifier: EPL-2.0 OR GPL-2.0 WITH Classpath-exception-2.0
  */
 
-package javax.security.auth.message.config;
+package jakarta.security.auth.message.config;
 
 import java.util.Map;
 import javax.security.auth.Subject;
-import javax.security.auth.message.AuthException;
+
+import jakarta.security.auth.message.AuthException;
 
 /**
  * This interface describes a configuration of ServerAuthConfiguration objects for a message layer and application
